+++ src/gui/gt_icons.h
@@ -15,6 +15,7 @@
 #include "gt_svgiconengine.h"
 
 #include <QIcon>
+#include <QUrl>
 
 namespace gt
 {
@@ -66,6 +67,22 @@ GT_GUI_EXPORT QIcon colorize(const QIcon& icon,
 
 namespace icon
 {
+
+/**
+ * @brief Returns the url of an icon
+ *
+ * Usage:
+ *
+ * auto iconurl = icon::url(icon::copy);
+ *
+ * @return
+ */
+template <typename IconGetter>
+QUrl url(IconGetter&& getIcon)
+{
+    return QString("qrc") + getIcon().name();
+}
+
 
 /**
  * @brief standardSizeSmall