+++ tests/modules/mdi_interface_ext/mdi/test_mdi_ext_viewer.cpp
@@ -10,8 +10,36 @@
 
 #include "test_mdi_ext_viewer.h"
 
+#include <gt_qmlaction.h>
+#include <gt_logging.h>
+
+#include <gt_icons.h>
+
 TestMdiExtViewer::TestMdiExtViewer()
 {
+    using namespace gt::gui;
+
+    cutAction = addToolbarAction("Cut", icon::url(icon::cut));
+    copyAction  = addToolbarAction("Copy", icon::url(icon::copy));
+    pasteAction = addToolbarAction("Paste", icon::url(icon::paste));
+
+    pasteAction->setEnabled(false);
+
+    connect(cutAction, &GtQmlAction::triggered, this, [this](){
+        gtInfo() << "Cut";
+        pasteAction->setEnabled(true);
+    });
+
+    connect(copyAction, &GtQmlAction::triggered, this, [this](){
+        gtInfo() << "Copy";
+        pasteAction->setEnabled(true);
+    });
+
+    connect(pasteAction, &GtQmlAction::triggered, this, [this](){
+        gtInfo() << "Paste";
+        pasteAction->setEnabled(false);
+    });
+
     setObjectName("Test Mdi Ext Viewer");
 }