+++ tests/modules/mdi_interface_ext/mdi/test_mdi_ext_viewer.h
@@ -34,6 +34,7 @@ public:
     bool allowsMultipleInstances() const override;
 
 private:
+    GtQmlAction *cutAction, *copyAction, *pasteAction;
 
 };